Update the db fields.

+++ database/migrations/2015_10_24_001613_add_geo_fields_users_table.php
@@ -14,8 +14,8 @@ class AddGeoFieldsUsersTable extends Migration
   {
     Schema::table('users', function ($table) {
       $table->string('address')->unique();
-      $table->float('lat');
-      $table->float('lng');
+      $table->decimal('lat', 11, 8);
+      $table->decimal('lng', 11, 8);
     });
   }
